requirements.txt
streamlit~=1.32.0
pandas~=2.2.3
plotly~=6.0.0
requests==2.31.0
Pillow~=10.0.0
numpy~=1.26.4
geopandas~=1.0.1
folium~=0.19.5
scikit-learn~=1.6.1
python-dotenv==1.0.0
shapely~=2.0.7